// common_components/src/interfaces/interfaces.ts
export interface UserInterface {
        passwordChangeReq: boolean;
        id: number;
        privileges: string[];
        isAuthenticated: boolean;
        token: string;
        userName: string;
        isSystemDefault: boolean;
        activeTab: number;
        values: Record<string, any>;
        errors: Record<string, any>;
        passwordReset: boolean;
        allowCancel: boolean;
        userDetails: UserDetails;
        userPreferences: UserPreferences;
        appType: string;
        platformType: string;
        localVMIP: string;
        zone: string;
        license: Record<string, any>;
    
}
// ...
export type Theme = 'light' | 'dark';

export interface UserDetails {
    id: number;
    username: string;
    fullName: string;
    password: string;
    email: string;
    description: string;
    isSystemDefault: boolean;
    isForcePasswordChange: boolean;
    role: {
        id: number;
        name: string;
        description: string;
        isSystemDefault: boolean;
        privileges: any | null; // Replace `any` with a more specific type if known
    };
}

export interface UserPreferences {
    userType: 'System' | 'Saml' | '';
    themePreference: 'dark' | 'light';
    username: string;
}